Added waves
Added the "waves" section to the config file so waves of units can be
created and easily reused. The "spawnwave" command was also added to
spawn a wave from its ID. I refactored how spawning works to make the
part of the program that handles spawning enemies handle queues of
enemies. This works for waves directly, and regular spawns use it by
generating a queue from the spawn command.