Added button class
The button class takes a position and a function that is called when the
button is clicked. It also accepts three images for the idle, mouseover,
and mousedown states. These images come from a basic surface generator
for now, but will eventually be basic text buttons (unless other images
are desired). A test button is on display on the main menu. Clicking
this button enters the game state on the test level. SSentities now
remove themselves when they reach either end of the path, a behavior
that will likely be changed later on.