Major Menu rework, and more
The main menu now leads to several incomplete pages for things like save
file selection and setting changes. The buttons are properly sized and
styled, and now have text labels. SSEntities now move along the path so
their feet (bottom center) are the origin of the image. I also changed
the knight image to be 36x36 as that is the size of the image itself
(the rest was empty space). I changed the TD background to be a somewhat
nicer grass one. Finally, I posed a question in one of the comments in
TD world that I am curious about.