Finished arrows, added a shorter test level
The arrow side of projectiles is now finished, with a few properties to
pull out to generalize the class to many kinds of projectiles. The new
test level features a much shorter path for faster testing, and is
closer to the final idea for the game. It also reveals the need for a
separate foreground layer that will be added in the future. The
"cfgloader" file was renamed to "loader" as it will also house the save
file loader.