Added foregrounds to SSworlds, refined TDworld buttons
[level]/fg.png is now loaded and drawn over the screen at the end of a
frame. The buttons for TDworld buttons have been swapped over to the
much prettier buttons from SSworlds. It looks a lot better now! I left
them spaced over to the right a bit so we can use the space to display
info like the cost of the action of the button that the user has their
mouse over.  The system for UI stuff on TD side is... honestly not
great. We should rework that ASAP before things get messier!