Further enhanced projectiles, added zombies, reorganized resources
Projectiles now base their arc height on the total flight distance and
stick to the specs they are given from their attack a bit more, though
some properties still do not work. Units can now define the size and
time of the frames in their sheet. As far as organization goes, I moved
all of the programmer art into a folder, deleted the old, not-so-great
test levels, and placed the player arrow projectile in the player
folder.