Added kings and the game over screen
Kings don't yet attack for some reason and the button on the game over
screen doesn't work, but it is a start! I also added the start of the
first skeleton level that Ben has been working on. Finally, I threw a
bunch of TODO's in that I will get to eventually.