Added bosses, started common image module
Bosses now follow their definition in config files and inform the game
when they die. To start optimizing the game, I added the "common"
module. To hopefully speed things up in the future (not that it is too
much of an issue at the moment), we should load all surfaces here and
keep copies of everything around so that we don't have to reload the
same image hundreds of times in a game. Right now, it holds a collection
of button sets, but I will have it load everything soon, including
images, fonts, and sounds when they are implemented.