Reworked attacks a bit
Units now stop moving while attacking for a set amount of time (which
will eventually be based on the length of their animation. Animations
will probably eventually find their home in the "sheetSet" module. I
moved a few files around to start working on that. SSentities have an
offset property to help position them on the path properly.