Added unlocks
Beating a level should unlock whatever that level states it will in the
config file, though it is completely untested. In the process, I added a
function to menus that adds the new unlocks to the save file. This
method calls another new one that makes sure level buttons are properly
enabled if their level is unlocked.