Added the pause menu
Clicking the "menu" button no longer exits the current game immediately,
but instead gives you the opportunity to either return to your game or
go to the level select screen. This menu will also eventually have a
settings button, but the settings menu will first need to have a way for
its back button to go to the pause menu rather than always going to the
main menu.