The monk button no longer crashes the game...
Instead, it spawns a monk! Unfortunately, due to how the unit spawning
buttons work I had to change every instance of the name "brother clint"
to "monk" to make the string one word. On the bright side, the true name
lives on in the name of the image, and he gets to be a battle monk until
I can get his actual "attack" implemented.